docs: 更新 README 和 TODO
- 在 README.zh-CN.md 中添加了对个人技术能力的简述
- 在 TODO.md 中增加了学习 AI 的计划
- 添加了新的图片资源 despise.png
