+++ README.zh-CN.md
@@ -9,8 +9,11 @@
 - 天赋，我是最没有的，但正因如此才不能放弃，也不能敷衍偷懒。
 
 PS:
-首先说下，对自己的总结：我愿意去学，但是好像没什么特别的天赋。很多东西知之甚少，缺乏对所涉及技术的深层理解。
+首先说下，对自己的总结：我愿意去学，但是好像没什么特别的天赋。
+很多东西知之甚少，缺乏对所涉及技术的深层理解。
 所以项目本身对于小白开发者可能是有用的，但没有什么很深度的笔记，大佬们就当乐子看好了。
+
+**简述：知其然，不知其所以然。**
 
 ---
