+++ log/TODO.md
@@ -25,3 +25,5 @@
 6. DeepSeek 的接入与学习
 
 7. 完成 gulp + electron-builder 打包任务
+
+8. 学习 AI,接受 AI,拥抱 AI
